fix: Moved rollDice to GameController.kt - restored deleted comments. Fixed bugs where the broadcast was not in the specific game.id but lobby-wide in the GameController.kt
